requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
certifi==2021.10.8
charset-normalizer==2.0.9; python_version >= '3'
coreapi==2.3.3
coreschema==0.0.4
django-filter==21.1
django==3.2.11
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3; python_version >= '3'
importlib-metadata==4.10.0; python_version < '3.10'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.0.3; python_version >= '3.6'
markdown==3.3.6
markupsafe==2.0.1; python_version >= '3.6'
packaging==21.3; python_version >= '3.6'
pillow==8.4.0
pyparsing==3.0.6; python_version >= '3.6'
python-slugify==5.0.2
pytz==2021.3
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ruamel.yaml.clib==0.2.6; python_version < '3.10' and platform_python_implementation == 'CPython'
ruamel.yaml==0.17.17; python_version >= '3'
sqlparse==0.4.2; python_version >= '3.5'
text-unidecode==1.3
tzdata==2021.5; sys_platform == 'win32'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
zipp==3.6.0; python_version >= '3.6'
